# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Crypto Bot API",
        "description": "API para gerenciamento de portfólios de criptomoedas, estratégias de trading e alertas",
        "contact": {
            "name": "Suporte Crypto Bot",
            "email": "support@cryptobot.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/v1/alerts": {
            "get": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Listar alertas",
                "description": "Listar todos os alertas do usuário autenticado.",
                "operationId": "9b0d3c8dafcdf1d483cfd0fc94c4253b",
                "parameters": [
                    {
                        "name": "portfolio_id",
                        "in": "query",
                        "description": "ID do portfólio para filtrar alertas",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "asset_id",
                        "in": "query",
                        "description": "ID do ativo para filtrar alertas",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Tipo de alerta (price, volume, technical)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "price",
                                "volume",
                                "technical"
                            ]
                        }
                    },
                    {
                        "name": "is_active",
                        "in": "query",
                        "description": "Filtrar por alertas ativos",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de alertas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "portfolio_id": {
                                                "type": "integer"
                                            },
                                            "asset_id": {
                                                "type": "integer"
                                            },
                                            "type": {
                                                "type": "string"
                                            },
                                            "condition": {
                                                "type": "string"
                                            },
                                            "value": {
                                                "type": "number"
                                            },
                                            "is_active": {
                                                "type": "boolean"
                                            },
                                            "last_triggered": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Criar alerta",
                "description": "Criar um novo alerta para um portfólio ou ativo do usuário autenticado.",
                "operationId": "7640e6ae7c632acc3067c5f938bdfaea",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "portfolio_id",
                                    "type",
                                    "condition",
                                    "value"
                                ],
                                "properties": {
                                    "portfolio_id": {
                                        "type": "integer"
                                    },
                                    "asset_id": {
                                        "type": "integer"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "price",
                                            "volume",
                                            "technical"
                                        ]
                                    },
                                    "condition": {
                                        "type": "string",
                                        "enum": [
                                            "above",
                                            "below",
                                            "equals"
                                        ]
                                    },
                                    "value": {
                                        "type": "number"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "notification_channels": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "cooldown_minutes": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Alerta criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "portfolio_id": {
                                            "type": "integer"
                                        },
                                        "asset_id": {
                                            "type": "integer"
                                        },
                                        "type": {
                                            "type": "string"
                                        },
                                        "condition": {
                                            "type": "string"
                                        },
                                        "value": {
                                            "type": "number"
                                        },
                                        "is_active": {
                                            "type": "boolean"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}